requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
charset-normalizer==3.2.0
discord-py-interactions==5.9.2
discord-typings==0.6.0
emoji==2.7.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
tomli==2.0.1
typing_extensions==4.7.1
yarl==1.9.2